+++ src/components/ButtonView/ButtonView.tsx
@@ -95,7 +95,7 @@ export const ButtonView: React.FC<Props> = ({
   if (!variable) {
     return (
       <Alert severity="info" title="Variable" data-testid={TEST_IDS.buttonView.noVariableMessage}>
-        Variable is not selected.
+        Variable is not selected. Constant, Data Source, Interval, AD hoc filters are not supported.
       </Alert>
     );
   }